music_platform/pyproject.toml
[tool.poetry]
name = "music_platform"
version = "0.1.0"
description = ""
authors = ["Mohammed Salah <mohamed.salah9972@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django-model-utils = "^4.2.0"
django-widget-tweaks = "^1.4.12"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
